fix tests; refactor app related utils
